[DATAHUBIO-38-chart-creation] Merged PlotData with Chart class
