+++ src/components/plot.jsx
@@ -2,17 +2,17 @@ import {CartesianGrid, Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAx
 import React from "react";
 import PropTypes from 'prop-types';
 
-function Plot({key, plotData, width, aspect}) {
+function Plot({key, chart, width, aspect}) {
     return (
         <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
             <ResponsiveContainer width={width} aspect={aspect}>
-                <LineChart data={plotData.points}>
+                <LineChart data={chart.points}>
                     <YAxis/>
                     <CartesianGrid stroke="#eee" strokeDasharray="5 5"/>
                     <XAxis dataKey={"timestamp"} hide={false}/>
                     <YAxis/>
                     {
-                        Object.entries(plotData.charts).map(chart => {
+                        Object.entries(chart.dataSeries).map(chart => {
                             return (<Line key={chart[0]}
                                           type="monotone"
                                           dataKey={"values." + chart[0]}
@@ -28,7 +28,7 @@ function Plot({key, plotData, width, aspect}) {
 
 Plot.propTypes = {
     key: PropTypes.string.isRequired,
-    plotData: PropTypes.string.isRequired,
+    chart: PropTypes.string.isRequired,
     width: PropTypes.number.isRequired,
     aspect: PropTypes.number.isRequired
 };
