+++ src/reducers/config.js
@@ -1,8 +1,8 @@
 import {LOAD_CONFIG} from '../actions/types'
-import {mockedPlotData} from "../utils/mockedData"
+import {mockedChart} from "../utils/mockedChart"
 
 const initialState = {
-    graphs: [mockedPlotData(), mockedPlotData(), mockedPlotData()]
+    graphs: [mockedChart(), mockedChart(), mockedChart()]
 }
 
 export default function configReducer(state = initialState, action) {
