+++ src/utils/chart.js
@@ -4,6 +4,7 @@ class Chart {
     metadata;
     viewingTimeInterval;
     data;
+    dataSeries = {};
 
     constructor(metadata) {
         this.metadata = metadata;
@@ -13,6 +14,20 @@ class Chart {
 
     addData(newData) {
         this.data.concat(newData);
+    }
+
+    addChartInfo(name, color = null) {
+        if (color == null) {
+            color = Chart.randomColor();
+        }
+        this.dataSeries[name] = {
+            "color": color
+        }
+    }
+
+    static randomColor() {
+        return "#" + Math.floor(Math.random() * 16777215)
+            .toString(16);
     }
 }
 
@@ -39,4 +54,4 @@ async function createCharts(metadataList) {
     return new ChartContainer(await Promise.all(dataLoaderCalls));
 }
 
-module.exports = {Chart, ChartContainer, createCharts}
+export {Chart, ChartContainer, createCharts}
